Try and move local tests into main tests dir. Rename local tests. Save test settings to unique json files to avoid race conditions
