+++ tests/GuiBaseTest.py
@@ -20,7 +20,7 @@ from onionshare_gui.mode.receive_mode import ReceiveMode
 
 class GuiBaseTest(object):
     @staticmethod
-    def set_up(test_settings):
+    def set_up(test_settings, settings_filename='/tmp/testsettings.json'):
         '''Create GUI with given settings'''
         # Create our test file
         testfile = open('/tmp/test.txt', 'w')
@@ -45,7 +45,6 @@ class GuiBaseTest(object):
         app = OnionShare(common, testonion, True, 0)
 
         web = Web(common, False, True)
-        settings_filename = '/tmp/testsettings.json'
         open(settings_filename, 'w').write(json.dumps(test_settings))
 
         gui = OnionShareGui(common, testonion, qtapp, app, ['/tmp/test.txt'], settings_filename, True)
