+++ tests/local_receive_mode_test.py
@@ -14,13 +14,13 @@ from onionshare_gui import *
 
 from .GuiBaseTest import GuiBaseTest
 
-class ShareModeStayOpenTest(unittest.TestCase):
+class ReceiveModeTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         test_settings = {
-            "close_after_first_download": False,
+            "receive_allow_receiver_shutdown": True
         }
-        cls.gui = GuiBaseTest.set_up(test_settings)
+        cls.gui = GuiBaseTest.set_up(test_settings, '/tmp/ReceiveModeTest.json')
 
     @classmethod
     def tearDownClass(cls):
@@ -32,7 +32,7 @@ class ShareModeStayOpenTest(unittest.TestCase):
 
     @pytest.mark.run(order=2)
     def test_run_all_share_mode_tests(self):
-        GuiBaseTest.run_all_share_mode_tests(self, False, True)
+        GuiBaseTest.run_all_receive_mode_tests(self, False, True)
 
 if __name__ == "__main__":
     unittest.main()
