+++ tests/local_share_mode_public_mode_test.py
@@ -14,14 +14,13 @@ from onionshare_gui import *
 
 from .GuiBaseTest import GuiBaseTest
 
-class ShareModeTimerTest(unittest.TestCase):
+class ShareModePublicModeTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         test_settings = {
-            "public_mode": False,
-            "shutdown_timeout": True,
+            "public_mode": True,
         }
-        cls.gui = GuiBaseTest.set_up(test_settings)
+        cls.gui = GuiBaseTest.set_up(test_settings, '/tmp/ShareModePublicModeTest.json')
 
     @classmethod
     def tearDownClass(cls):
@@ -32,8 +31,8 @@ class ShareModeTimerTest(unittest.TestCase):
         GuiBaseTest.run_all_common_setup_tests(self)
 
     @pytest.mark.run(order=2)
-    def test_run_all_share_mode_timer_tests(self):
-        GuiBaseTest.run_all_share_mode_timer_tests(self, False)
+    def test_run_all_share_mode_tests(self):
+        GuiBaseTest.run_all_share_mode_tests(self, True, False)
 
 if __name__ == "__main__":
     unittest.main()
