+++ tests/local_share_mode_stay_open_test.py
@@ -14,14 +14,13 @@ from onionshare_gui import *
 
 from .GuiBaseTest import GuiBaseTest
 
-class ReceiveModePublicModeTest(unittest.TestCase):
+class ShareModeStayOpenTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         test_settings = {
-            "public_mode": True,
-            "receive_allow_receiver_shutdown": True
+            "close_after_first_download": False,
         }
-        cls.gui = GuiBaseTest.set_up(test_settings)
+        cls.gui = GuiBaseTest.set_up(test_settings, '/tmp/ShareModeStayOpenTest.json')
 
     @classmethod
     def tearDownClass(cls):
@@ -33,7 +32,7 @@ class ReceiveModePublicModeTest(unittest.TestCase):
 
     @pytest.mark.run(order=2)
     def test_run_all_share_mode_tests(self):
-        GuiBaseTest.run_all_receive_mode_tests(self, True, True)
+        GuiBaseTest.run_all_share_mode_tests(self, False, True)
 
 if __name__ == "__main__":
     unittest.main()
