+++ tests/local_share_mode_test.py
@@ -12,18 +12,14 @@ from onionshare.web import Web
 from onionshare import onion, strings
 from onionshare_gui import *
 
-from .GuiBaseTest import GuiBaseTest
+from .GuiBaseTest import GuiBaseTest 
 
-class ShareModePersistentSlugTest(unittest.TestCase):
+class ShareModeTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         test_settings = {
-            "public_mode": False,
-            "slug": "",
-            "save_private_key": True,
-            "close_after_first_download": False,
         }
-        cls.gui = GuiBaseTest.set_up(test_settings)
+        cls.gui = GuiBaseTest.set_up(test_settings, '/tmp/ShareModeTest.json')
 
     @classmethod
     def tearDownClass(cls):
@@ -35,14 +31,7 @@ class ShareModePersistentSlugTest(unittest.TestCase):
 
     @pytest.mark.run(order=2)
     def test_run_all_share_mode_tests(self):
-        GuiBaseTest.run_all_share_mode_tests(self, False, True)
-        global slug
-        slug = self.gui.share_mode.server_status.web.slug
-
-    @pytest.mark.run(order=3)
-    def test_have_same_slug(self):
-        '''Test that we have the same slug'''
-        self.assertEqual(self.gui.share_mode.server_status.web.slug, slug)
+        GuiBaseTest.run_all_share_mode_tests(self, False, False)
 
 if __name__ == "__main__":
     unittest.main()
