+++ tests/local_share_mode_timer_test.py
@@ -12,14 +12,16 @@ from onionshare.web import Web
 from onionshare import onion, strings
 from onionshare_gui import *
 
-from .GuiBaseTest import GuiBaseTest 
+from .GuiBaseTest import GuiBaseTest
 
-class ShareModeTest(unittest.TestCase):
+class ShareModeTimerTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         test_settings = {
+            "public_mode": False,
+            "shutdown_timeout": True,
         }
-        cls.gui = GuiBaseTest.set_up(test_settings)
+        cls.gui = GuiBaseTest.set_up(test_settings, '/tmp/ShareModeTimerTest.json')
 
     @classmethod
     def tearDownClass(cls):
@@ -30,8 +32,8 @@ class ShareModeTest(unittest.TestCase):
         GuiBaseTest.run_all_common_setup_tests(self)
 
     @pytest.mark.run(order=2)
-    def test_run_all_share_mode_tests(self):
-        GuiBaseTest.run_all_share_mode_tests(self, False, False)
+    def test_run_all_share_mode_timer_tests(self):
+        GuiBaseTest.run_all_share_mode_timer_tests(self, False)
 
 if __name__ == "__main__":
     unittest.main()
