--- tests_gui_local/__init__.py
@@ -1,1 +1,0 @@
-from .GuiBaseTest import GuiBaseTest
